fs.rs,ops.rs,block/*: Split fs and block structs into traits than can be trait objects.
I really like the fact that the rust generics can monomorph down to highly
optimized code paths, but unfortunately the more generic structure you have
nested, the harder it is to deal with at the top layer (cli code in our
case). It's flat out impossible to keep a list of open images in a vector,
for example. The only way to do it is to have trait objects, but then you
give up the monomorphization.

I didn't want to close off that option completely so there compromises
here. Creating a new member function means putting the definition in a
trait--this isn't so bad, it's akin to adding a class member function to a
c++ header file. But it _also_ needs to go into the Box impl, since Rust
doesn't auto implement traits on boxed trait objects (which is kind of
annoying). So RT11FS and PhysicalBlockDevice are now traits that can be
trait objects and there are impls for each as boxes. They are also structs
that can be used in the generic monomorphic way (I'm not willing to give
that up, even though I have no use for it in this code).

This patch also converts ops.rs to use the trait objects, which makes it
easier to deal with at the higher levels.