pdpfs.js: Convert all closures in useSelection() to useCallback()
I'm not sure I like this change. React stuff always feels a little
hacky. Not sure how to organize all the sub functions for this though.

Perhaps a class (with a cheap constructor) that gets reinstantiated in a
useEffect()? Maybe just the spanlist could be a class? The function is
unwieldy even after extracting from diskimageview...