pdpfs.js: Stop editing before we sent the mv off to fix error dialog being wonky
If there was an error the app would pop up a dialog. This would cause the
text entry to blur which would call its onBlur handler causing another
rename and a 2nd error. That 2nd error would kill the first dialog and bring
up another identicaly one. It was real ugly.